Do not resume yielded RoCs that are removed from the list while iterating
When the YieldedRequestOnConn's event fires in epoll, a snapshot of the
queue of RoCs to be resumed is made (the active/inactive queues are
swapped). This snapshot is then iterated over and each RoC in it is
resumed. The problem is that it's possible for RoCs to be removed from the
*active* queue while an iteration is in progress. This means that the RoC
is no longer waiting to be resumed. The iteration must take this into
account and not resume any RoCs that have been removed from the queue.

NOTE: this fix is a quick hack. It might be better to use a different data
structure internally (i.e. not a TreeQueue).